Implemented DecrementLife method (Basic Method for attacking a Player) and the Constructor for the Player Attribute "position" (Player) fixed (two dimension array has a different expression in C# ;)) Wrote tests for setting the life of Monsters and the Hero and for the DecrementLife method (Attack Player)

+++ Backend/CommandPatternLauncher/src/MDC/Gamedata/Command.cs
@@ -72,12 +72,11 @@ namespace MDC.Gamedata
         {
             //TODO: Throw Exception if AttackedPlayer not reachable 
 
-            int attackBoost = SourcePlayer.AttackBoost; 
+            double attackBoost = SourcePlayer.AttackBoost; 
             CharacterType characterType = SourcePlayer.CharacterType;
         
             TargetPlayer.DecrementLife(attackBoost, characterType); 
-            
-            
+                        
         }
     }
